requirements.txt
fastmcp>=2.0.0
pydantic>=2.0
uvicorn>=0.30.0
python-dotenv>=1.0.0